+++ tests/unit/server-helpers.spec.ts
@@ -6,6 +6,7 @@ import {
   createUploadFailureHandler,
   createUploadLogger,
   createUploadParsers,
+  makeChildLoggerExtractor,
   moveRaw,
 } from "../../src/server-helpers";
 import { describe, expect, test, vi } from "vitest";
@@ -22,43 +23,36 @@ import createHttpError from "http-errors";
 describe("Server helpers", () => {
   describe("createParserFailureHandler()", () => {
     test("the handler should call next if there is no error", () => {
-      const rootLogger = makeLoggerMock();
       const handler = createParserFailureHandler({
         errorHandler: defaultResultHandler,
-        rootLogger,
+        getChildLogger: () => makeLoggerMock(),
       });
       const next = vi.fn();
       handler(undefined, makeRequestMock(), makeResponseMock(), next);
       expect(next).toHaveBeenCalledTimes(1);
     });
 
-    test("the handler should call error handler with a child logger", async () => {
+    test("the handler should call error handler", async () => {
       const errorHandler = new ResultHandler({
         positive: vi.fn(),
         negative: vi.fn(),
         handler: vi.fn(),
       });
       const spy = vi.spyOn(errorHandler, "execute");
-      const rootLogger = makeLoggerMock({ isRoot: true });
       const handler = createParserFailureHandler({
         errorHandler,
-        rootLogger,
+        getChildLogger: () => makeLoggerMock(),
       });
       await handler(
         new SyntaxError("Unexpected end of JSON input"),
         makeRequestMock(),
-        makeResponseMock({
-          locals: {
-            [metaSymbol]: { logger: makeLoggerMock({ isChild: true }) },
-          },
-        }),
+        makeResponseMock(),
         vi.fn<any>(),
       );
       expect(spy).toHaveBeenCalledTimes(1);
       expect(spy.mock.calls[0][0].error).toEqual(
         createHttpError(400, "Unexpected end of JSON input"),
       );
-      expect(spy.mock.calls[0][0].logger).toHaveProperty("isChild", true);
     });
   });
 
@@ -70,10 +64,9 @@ describe("Server helpers", () => {
         handler: vi.fn(),
       });
       const spy = vi.spyOn(errorHandler, "execute");
-      const rootLogger = makeLoggerMock({ isRoot: true });
       const handler = createNotFoundHandler({
         errorHandler,
-        rootLogger,
+        getChildLogger: () => makeLoggerMock(),
       });
       const next = vi.fn();
       const requestMock = makeRequestMock({
@@ -81,17 +74,11 @@ describe("Server helpers", () => {
         path: "/v1/test",
         body: { n: 453 },
       });
-      const responseMock = makeResponseMock({
-        locals: {
-          [metaSymbol]: { logger: makeLoggerMock({ isChild: true }) },
-        },
-      });
+      const responseMock = makeResponseMock();
       await handler(requestMock, responseMock, next);
       expect(next).toHaveBeenCalledTimes(0);
       expect(spy).toHaveBeenCalledTimes(1);
       expect(spy.mock.calls[0]).toHaveLength(1);
-      expect(spy.mock.calls[0][0]).toHaveProperty("logger");
-      expect(spy.mock.calls[0][0].logger).toHaveProperty("isChild", true);
       expect(spy.mock.calls[0][0].error).toEqual(
         createHttpError(404, "Can not POST /v1/test"),
       );
@@ -102,7 +89,6 @@ describe("Server helpers", () => {
     });
 
     test("should call Last Resort Handler in case of ResultHandler is faulty", () => {
-      const rootLogger = makeLoggerMock();
       const errorHandler = new ResultHandler({
         positive: vi.fn(),
         negative: vi.fn(),
@@ -111,7 +97,7 @@ describe("Server helpers", () => {
       const spy = vi.spyOn(errorHandler, "execute");
       const handler = createNotFoundHandler({
         errorHandler,
-        rootLogger,
+        getChildLogger: () => makeLoggerMock(),
       });
       const next = vi.fn();
       const requestMock = makeRequestMock({
@@ -170,7 +156,7 @@ describe("Server helpers", () => {
   });
 
   describe("createUploadParsers()", async () => {
-    const rootLogger = makeLoggerMock();
+    const loggerMock = makeLoggerMock();
     const beforeUploadMock = vi.fn();
     const parsers = await createUploadParsers({
       config: {
@@ -183,9 +169,9 @@ describe("Server helpers", () => {
           },
         },
         cors: false,
-        logger: rootLogger,
+        logger: { level: "silent" },
       },
-      rootLogger,
+      getChildLogger: () => loggerMock,
     });
     const requestMock = makeRequestMock();
     const responseMock = makeResponseMock();
@@ -194,7 +180,7 @@ describe("Server helpers", () => {
     test("should return an array of RequestHandler", () => {
       expect(parsers).toEqual([
         expect.any(Function), // uploader with logger
-        expect.any(Function), // createUploadFailueHandler()
+        expect.any(Function), // createUploadFailureHandler()
       ]);
     });
 
@@ -208,12 +194,12 @@ describe("Server helpers", () => {
     });
 
     test("should install the uploader with its special logger", async () => {
-      const interalMw = vi.fn();
-      fileUploadMock.mockImplementationOnce(() => interalMw);
+      const internalMw = vi.fn();
+      fileUploadMock.mockImplementationOnce(() => internalMw);
       await parsers[0](requestMock, responseMock, nextMock);
       expect(beforeUploadMock).toHaveBeenCalledWith({
         request: requestMock,
-        logger: rootLogger,
+        logger: loggerMock,
       });
       expect(fileUploadMock).toHaveBeenCalledTimes(1);
       expect(fileUploadMock).toHaveBeenCalledWith({
@@ -223,7 +209,7 @@ describe("Server helpers", () => {
         limits: { fileSize: 1024 },
         logger: { log: expect.any(Function) }, // @see createUploadLogger test
       });
-      expect(interalMw).toHaveBeenCalledWith(
+      expect(internalMw).toHaveBeenCalledWith(
         requestMock,
         responseMock,
         nextMock,
@@ -242,6 +228,27 @@ describe("Server helpers", () => {
       moveRaw(requestMock, makeResponseMock(), nextMock);
       expect(requestMock.body).toEqual({ raw: buffer });
       expect(nextMock).toHaveBeenCalled();
+    });
+  });
+
+  describe("makeChildLoggerExtractor()", () => {
+    const rootLogger = makeLoggerMock();
+    const getChildLogger = makeChildLoggerExtractor(rootLogger);
+
+    test("should extract child logger from request", () => {
+      const request = makeRequestMock({
+        res: {
+          locals: {
+            [metaSymbol]: { logger: makeLoggerMock({ isChild: true }) },
+          },
+        },
+      });
+      expect(getChildLogger(request)).toHaveProperty("isChild", true);
+    });
+
+    test("should fall back to root", () => {
+      const request = makeRequestMock();
+      expect(getChildLogger(request)).toEqual(rootLogger);
     });
   });
 });